grpc-server/requirements.txt
grpcio==1.62.1
grpcio-tools==1.62.1
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pytest==7.4.3
pytest-mock==3.12.0